% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{Function1}
\alias{Function1}
\title{Summary Statistics for Flow_sample data set}
\usage{
Function1(data, category, values)
}
\value{
A tibble with the summary statistic information
}
\description{
This function outputs minimum, maximum, and mean summary statistics for flow measurements for each recorded month in the flow_samples data set. The function identifies the data, groups by a given category, and calculates summary statistics for the measured values.
}
