% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{Function1}
\alias{Function1}
\title{Summary Statistics}
\usage{
Function1(data, category, values)
}
\arguments{
\item{data}{The data set to run summary statistics on}

\item{category}{The defined category that statistics are calculated within}

\item{values}{The quantitative data values to run statistics on}
}
\value{
A tibble with the summary statistic information
}
\description{
This function outputs minimum, maximum, and mean summary statistics for quantitative measurements for defined categories in the data set.
}
